MWPW-156612 fragment merch-card variant layouts (#2810)
* MWPW-156612 fragment merch-card variant layouts

- move from merch-card.js code each layout code related to a given variant under variants/<variant> class
- move from global.css.js css related to a given variant under variants/<variant>.css.js
- move from merch-card.css.js wc css related to a given variant under variants/<variant>.variantStyle
- uses variants/variants as an index to get related class from related variant, and wc style too,
- uses variants/VariantLayout, superclass of all variants, to insert css if card is used, and hold common code

* MWPW-156612 changing title selector

to something independent of the layout (not initialized when used at first in collections)

* MWPW-156612 forking style insertion

* MWPW-156612 styles fixes

* MWPW-156612 review comments

* MWPW-156612 fixing unit tests

* MWPW-156612 address review comments

* MWPW-156612 catalog & image test cov

* MWPW-156612 100% coverage for variants

* MWPW-156612 fixing (again) ut

* MWPW-156612 fix one inline heading style

* MWPW-156612 bg img height
