+++ libs/features/mas/web-components/src/variants/twp.js
@@ -1,0 +1,67 @@
+import { VariantLayout } from "./variant-layout";
+import { html, css } from 'lit';
+import { CSS } from './twp.css.js';
+
+export class TWP extends VariantLayout {
+  constructor(card) {
+    super(card);
+  }
+
+  getGlobalCSS() {
+    return CSS;
+  }
+
+  renderLayout() {
+    return html`${this.badge}
+      <div class="top-section">
+          <slot name="icons"></slot>
+          <slot name="heading-xs"></slot>
+          <slot name="body-xs-top"></slot>
+      </div>
+      <div class="body">
+          <slot name="body-xs"></slot>
+      </div>
+      <footer><slot name="footer"></slot></footer>`;
+  }
+
+  static variantStyle = css`
+    :host([variant='twp']) {
+      padding: 4px 10px 5px 10px;
+    }
+    .twp-badge {
+      padding: 4px 10px 5px 10px;
+    }
+
+    :host([variant='twp']) ::slotted(merch-offer-select) {
+      display: none;
+    }
+
+    :host([variant='twp']) .top-section {
+      flex: 0;
+      display: flex;
+      flex-direction: column;
+      justify-content: flex-start;
+      height: 100%;
+      gap: var(--consonant-merch-spacing-xxs);
+      padding: var(--consonant-merch-spacing-xs)
+          var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-xs)
+          var(--consonant-merch-spacing-xs);
+      align-items: flex-start;
+    }
+
+    :host([variant='twp']) .body {
+      padding: 0 var(--consonant-merch-spacing-xs);
+    }
+    
+    :host([aria-selected]) .twp-badge {
+        margin-inline-end: 2px;
+        padding-inline-end: 9px;
+    }
+
+    :host([variant='twp']) footer {
+      gap: var(--consonant-merch-spacing-xxs);
+      flex-direction: column;
+      align-self: flex-start;
+    }
+  `;
+}
