+++ libs/features/mas/web-components/test/merch-card.mini-compare.mobile.test.html
@@ -53,7 +53,7 @@
 </head>
 
 <body>
-    <script type="module" src="./merch-card.mini-compare.test.html.js"></script>
+    <script type="module" src="./merch-card.mini-compare.mobile.test.html.js"></script>
     <main>
         <div id="sidenav">
             <button onclick="toggleRTL()">Toggle RTL</button>
